hosted , payment grpah setup finished

+++ server/controllers/auth.controller.js
@@ -4,19 +4,32 @@ const Otp = require("../model/otp");
 const bcrypt = require("bcrypt");
 const transporter  = require("../config/nodemailer");
 
-require("dotenv").config();
+
+require("dotenv").config()
+
 
 const generateOtp = () =>
   Math.floor(100000 + Math.random() * 900000).toString();
 
 exports.sendOtp = async (req, res) => {
-  const { email } = req.body;
-  console.log(process.env.EMAIL_USER, process.env.EMAIL_PASS);
-  console.log(email);
-
-  const otp = generateOtp();
+  const { email ,username} = req.body;
 
   try {
+
+    const existingEmail = await Host.findOne({ email });
+
+    const existingName = await Host.findOne({ username });
+
+ 
+    if (existingName) {
+      return res.status(403).json({ message: "Username already exists" });
+    }
+    if (existingEmail) {
+      return res.status(403).json({ message: "Email already exists" });
+    }
+
+    const otp = generateOtp();
+
     await Otp.deleteMany({ email });
 
     await Otp.create({ email, otp });
@@ -36,7 +49,7 @@ exports.sendOtp = async (req, res) => {
 
 exports.verifyOtp = async (req, res) => {
   const { username, email, password, otp } = req.body;
-  console.log(req.body);
+
 
   try {
     const otpDoc = await Otp.findOne({ email, otp });
@@ -74,7 +87,14 @@ exports.userlogin = async (req, res) => {
       return res.status(400).send("Email and password are required");
     }
 
-    const user = await Host.findOne({ email });
+    const user = await Host.findOne({ email }).select('+password');
+
+
+    // if (user.isResticted)
+    //   return res
+    //     .status(403)
+    //     .json({ message: "Your account is restricted from logging in." });
+
 
     if (!user) {
       return res.status(404).send("User not found");
@@ -83,7 +103,6 @@ exports.userlogin = async (req, res) => {
     if (!user.password || !(await bcrypt.compare(password, user.password))) {
       return res.status(404).send("The password is incorrect");
     }
-    console.log('after pass');
     
 
     const token = jwt.sign({ email: user.email, userId: user._id }, process.env.JWT_SECRET, { expiresIn: '1d' });
@@ -145,8 +164,6 @@ exports.updateprofile = async (req, res) => {
     user.phone = phone || user.phone;
     user.street = street || user.street;
     user.gender = gender || user.gender;
-    
-    user.profileSetup = true;
 
     await user.save();
